converted all tasks to zappa tasks
