+++ intake/services/submissions.py
@@ -46,7 +46,7 @@ def send_to_newapps_bundle_if_needed(submission, organizations):
     sf_ids = [org.id for org in organizations if org.slug == 'sf_pubdef']
     if sf_ids:
         app = submission.applications.filter(organization_id=sf_ids[0]).first()
-        tasks.add_application_pdfs.delay(app.id)
+        tasks.add_application_pdfs(app.id)
 
 
 def fill_pdfs_for_submission(submission, organizations=None):
